feat: run git fetch before deployment steps to allow user to git checkout
also did a refactor of moving all convenience code into it's own module. I did this because some tests broke when all convenience code was in lib/steps/deploy.ts and the tests started to get complex when they didn't need to.
