+++ lib/steps/deploy.test.ts
@@ -1,11 +1,9 @@
 import { assertEquals, assertRejects } from "@std/assert"
-import { afterEach, before, beforeEach, describe, it } from "@std/testing/bdd"
+import { afterEach, describe, it } from "@std/testing/bdd"
 import { restore, stub } from "@std/testing/mock"
 import { exec } from "../exec.ts"
 import { DeployStepImpl } from "./deploy.ts"
 import { DeployStepInput } from "../types/environment.ts"
-import { mock, when } from "../mock/mock.ts"
-import { GitHubActions } from "../github-actions.ts"
 
 const defaultEnvironment: DeployStepInput = {
   gitCurrentBranch: "main",
@@ -18,13 +16,6 @@ const defaultEnvironment: DeployStepInput = {
 }
 
 describe("run the user given deploy commands", () => {
-  let githubActionsMock = mock<GitHubActions>()
-
-  beforeEach(() => {
-    githubActionsMock = mock<GitHubActions>()
-    when(githubActionsMock, "getGitConfigInput", () => undefined)
-  })
-
   afterEach(() => {
     restore()
   })
@@ -42,7 +33,7 @@ describe("run the user given deploy commands", () => {
 
     Deno.env.set("INPUT_DEPLOY", command)
 
-    await new DeployStepImpl(exec, githubActionsMock).runDeploymentCommands({
+    await new DeployStepImpl(exec).runDeploymentCommands({
       environment: defaultEnvironment,
     })
 
@@ -60,7 +51,7 @@ describe("run the user given deploy commands", () => {
 
     Deno.env.set("INPUT_DEPLOY", "")
 
-    await new DeployStepImpl(exec, githubActionsMock).runDeploymentCommands({
+    await new DeployStepImpl(exec).runDeploymentCommands({
       environment: defaultEnvironment,
     })
 
@@ -81,7 +72,7 @@ describe("run the user given deploy commands", () => {
     Deno.env.set("INPUT_DEPLOY", command)
 
     await assertRejects(async () => {
-      await new DeployStepImpl(exec, githubActionsMock).runDeploymentCommands({
+      await new DeployStepImpl(exec).runDeploymentCommands({
         environment: defaultEnvironment,
       })
     })
@@ -99,7 +90,7 @@ describe("run the user given deploy commands", () => {
       }
     })
 
-    await new DeployStepImpl(exec, githubActionsMock).runDeploymentCommands({
+    await new DeployStepImpl(exec).runDeploymentCommands({
       environment: defaultEnvironment,
     })
   })
